BUILD Add Heading UI Component
Built out the Heading UI component for setting headings with
sizing options decoupled from the tag. Also seen the issue for the
icon sizing type on the SocialIcons sections component.

In the future we plan on installing PostCSS to use the
postcss-calc integration to set a base size. Then multiply the
base heading size for each heading size. This way it is easy to
change all the headings bigger or smaller by editing the base size.